pyproject.toml
[tool.poetry]
name = "gtm-scraper"
version = "0.1.0"
description = ""
authors = ["James Spillmann <james.spillmann@surfside.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httplib2 = "^0.22.0"
apiclient = "^1.0.4"
oauth2client = "^4.1.3"
google-api-python-client = "^2.159.0"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
aiohttp = "^3.11.11"
tqdm = "^4.67.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
scape = 'gtm_scraper.run'